+++ public_html/app/Models/Informe.php
@@ -10,5 +10,6 @@ class Informe extends Model {
     protected $fillable = [
         'texto',
         'visto',
+        'endereco',
     ];
 }